refact: rename Editor to StudyPane
