+++ src/components/StudyPane/Header/Tabs.tsx
