+++ src/components/StudyPane/Header/Title.tsx
