+++ src/components/StudyPane/Header/index.tsx
